requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiofiles==23.2.1
Flake8==7.0.0
black==24.1.1
pytest==8.0.0
pytest-asyncio==0.23.5 